# [Java.Interop.Tools.Cecil] use File.Exists instead of DirectoryGetFile