@jonathanpeppers 2020-03-09
I was puzzled at how this simple change helps so much, but it looks like after I add some logging `SearchDirectory` is called ~3,432 times for a Xamarin.Forms app:

https://github.com/xamarin/java.interop/blob/27cfd452c6e543374728ace77af05ab65f9952b6/src/Java.Interop.Tools.Cecil/Java.Interop.Tools.Cecil/DirectoryAssemblyResolver.cs#L252

In a first build, both `LinkAssembliesNoShrink` and `GenerateJavaStubs` call this for every assembly `*` every search directory (until found) `*` 2 MSBuild tasks * 2 `.dll` vs `.exe`.